# dist/swagger/swagger.json
{
    "components": {
...
        "schemas": {
...
            "IPortofolioData": {
                "properties": {
                    "apps": {
                        "items": {
                            "$ref": "#/components/schemas/IApp"
                        },
                        "type": "array"
                    },
                    "apis": {
                        "items": {
                            "$ref": "#/components/schemas/IApiRoute"
                        },
                        "type": "array"
                    },
                    "building": {
                        "items": {
                            "$ref": "#/components/schemas/IBosData"
                        },
                        "type": "array"
                    }
                },
                "type": "object",
                "additionalProperties": {}
            },
...
            "IPortofolioInfo": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "appIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "apiIds": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "info": {
        "title": "BOS config admin API routes",
        "version": "1.0.0",
        "description": "Welcome to the reference documentation for Portofolio asset manager (PAM) REST API! </br></br>To learn about the common use cases and concept this REST APIs, check out our resource center  (<a>https://resourcecenter.fr.spinalcom.com/ressources-développeur</a>).</br></br>In addition to the API Reference, we also provide complete documentation for adminnistration API, that need to install and implement Spinalcore Digital Twin Operating System on their built environment.",
        "license": {
            "name": "ISC"
        },
        "contact": {
            "name": "Spinalcom",
            "email": "developers@spinalcom.com"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/pam/add_portofolio": {
            "post": {
                "operationId": "AddPortofolio",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IPortofolioData"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IPortofolioInfo"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/update_portofolio/{portofolioId}": {
            "put": {
                "operationId": "UpdatePortofolio",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IPortofolioData"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "portofolioId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IEditPortofolio"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/rename_portofolio/{id}": {
            "put": {
                "operationId": "RenamePortofolio",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ],
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pam/get_all_portofolio": {
            "get": {
                "operationId": "GetAllPortofolio",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": {},
                                            "type": "array"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/v1/pam/get_portofolio/{id}": {
            "get": {
                "operationId": "GetPortofolio",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {},
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/v1/pam/get_portofolio_details/{id}": {
            "get": {
                "operationId": "GetPortofolioDetails",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "$ref": "#/components/schemas/IPortofolioData"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "message"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Portofolio"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}